Drop down camera lengths
Changed makeCameraDropDownMenu() function in toolkit
Added getAvailableCameraLengths() to dataObject
Added check to find software version in dataObject when it is built and
a function to return this to other functions.
Note: This breaks the Camera Length list because it does not update if
the tag group is changed AFTER the dialog has been made. But that is a
different problem.