CL Drop Down Menu
Re-ordered the bootup order so that persistent data would be loaded
before the toolkit was displayed. The camera length drop down menu will
now include all of the persistent values it needs.
Note: Toolkit will need to restart if the CL values are customized.
Note 2: The function to load scales has not been editted to use the new
tag system. It works by coincidence.