Untested Persistent tag changes
Added isImagingMode and isDiffractionMode to the camera control class.
These will check if the microscope is in the correct mode and return
0/1.
Changed takeDPImage() to add persistent image tags to its returned
image.
Added saveImageInImageSet to the toolkit. This uses an image's own
persistent tags to find its image set and save the file with the correct
title and in its own subdirectory.