Image Save Checks
Moved SaveImageInImageSet to the Image Set Tools.
Split it into two functions.
generateFileNameForImageInImageSet() generates the filename and checks
to see if it can save in the right directory.
The second one saves the file using the first function and also updates
the image set data with the filename.

Added identifyImageInImageSet to ImageSetTools. This gets information
from an images persistent tags so it can be found in an imageSet.

Added getImageDataFromImageSet. Gets an image's specific tag group from
an imageSet using a few identifying pieces of information.